todo/requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.3.0
click==8.1.3
Django==4.0.4
djangorestframework==3.13.1
more-itertools==8.13.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pip==22.1.1
platformdirs==2.5.2
pluggy==0.13.1
py==1.11.0
pyparsing==3.0.9
pytest==5.4.3
pytz==2022.1
setuptools==60.6.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
wcwidth==0.2.5
wheel==0.37.1
